Fix import issue where not preloading, grid tweaks
