+++ src/mainwindow.cpp
@@ -288,7 +288,6 @@ MainWindow::MainWindow(QWidget *parent) : QMainWindow(parent), ui(new Ui::MainWi
 
     connect(pmContainer, SIGNAL(fileToOpen(QString, bool)), SLOT(openProject(QString, bool)));
     connect(pmContainer, SIGNAL(fileToCreate(QString, QString)), SLOT(newProject(QString, QString)));
-    connect(pmContainer, SIGNAL(importProject()), SLOT(importSceneFromZip()));
     connect(pmContainer, SIGNAL(exportProject()), SLOT(exportSceneAsZip()));
 
     // toolbar stuff
@@ -1367,83 +1366,6 @@ bool MainWindow::isModelExtension(QString extension)
        extension == "dae")
         return true;
     return false;
-}
-
-QString importProjectName;
-QStringList fileNames;
-int on_extract_entry(const char *filename, void *arg) {
-    QFileInfo fInfo(filename);
-    if (fInfo.suffix() == "db") {
-        importProjectName = fInfo.baseName();
-    } else {
-        fileNames.append(filename);
-    }
-
-    return 0;
-}
-
-void MainWindow::importSceneFromZip()
-{
-    auto filename = QFileDialog::getOpenFileName(this,      "Import World",
-                                                 nullptr,   "Jahshaka Project (*.zip)");
-
-    if (filename.isEmpty() || filename.isNull()) return;
-
-    // get the current project working directory
-    auto pFldr = IrisUtils::join(QStandardPaths::writableLocation(QStandardPaths::DocumentsLocation),
-                                 Constants::PROJECT_FOLDER);
-    auto defaultProjectDirectory = settings->getValue("default_directory", pFldr).toString();
-
-    // create a temporary directory and extract our project into it
-    // we need a sure way to get the project name, so we have to extract it first and check the blob
-    QTemporaryDir temporaryDir;
-    temporaryDir.setAutoRemove(false);
-    if (temporaryDir.isValid()) {
-        zip_extract(filename.toStdString().c_str(),
-                    temporaryDir.path().toStdString().c_str(),
-                    on_extract_entry,
-                    Q_NULLPTR);
-    }
-
-    // now extract the project to the default projects directory with the name
-    if (!importProjectName.isEmpty() || !importProjectName.isNull()) {
-        auto pDir = QDir(defaultProjectDirectory).filePath(importProjectName);
-        QDir workingTempDirectory(temporaryDir.path());
-
-        QDir dirMaker;
-        dirMaker.mkdir(pDir);
-
-        struct zip_t *zip = zip_open(filename.toStdString().c_str(), 0, 'r');
-
-        for (int i = 0; i < fileNames.count(); i++) {
-            QFileInfo fInfo(fileNames[i]);
-            auto file = QString(workingTempDirectory.relativeFilePath(fileNames[i])).toStdString().c_str();
-
-            // we need to pay special attention to directories since we want to create empty ones as well
-            // also directories need to exist prior to a file being written
-            if (fInfo.isDir()) {
-                dirMaker.mkdir(QDir(pDir).filePath(file));
-            }
-            else {
-                zip_entry_open(zip, file);
-                zip_entry_fread(zip, QDir(pDir).filePath(file).toStdString().c_str());
-            }
-
-            // we close each entry after a successful write
-            zip_entry_close(zip);
-        }
-
-        zip_close(zip);
-
-        auto open = db->importProject(QDir(temporaryDir.path()).filePath(importProjectName));
-        if (open) {
-            Globals::project->setProjectPath(pDir);
-            pmContainer->hide();
-            openProject(pDir);
-        }
-    }
-
-    temporaryDir.remove();
 }
 
 void MainWindow::exportSceneAsZip()
