+++ src/mainwindow.h
@@ -172,7 +172,6 @@ private:
     bool isModelExtension(QString extension);
 
 public slots:
-    void importSceneFromZip();
     void exportSceneAsZip();
 
     //scenegraph
